introduced CFP Preferences controller, shared between CFP comitee and GT
