+++ conf/messages.fr
@@ -848,7 +848,7 @@ title.maxLength=125 caractères max.
 
 demoLevel.help=Slides, atelier, ou démonstration en direct... à vous de préciser. Cette information n''est utile que pour le comité de sélection.
 
-email.notifications.label=Notifications par email (CFP seulement)
+email.notifications.label=Notifications par email
 email.notifications.autowatch.title=Autowatch
 email.notifications.autowatch.desc1=L''autowatch permet de s''abonner à des notifications automatiquement à partir d''un certain évènement sur la proposition.
 email.notifications.autowatch.desc2=Une fois observateur d''une proposition, vous recevrez des notifications email lors de la survenue d''évènements sur la proposition (cf paramétrage dans la section suivante)
