+++ conf/routes
@@ -41,7 +41,6 @@ GET           /cfp/newSpeakerForExistingWebuser                                 
 
 POST          /cfp/profile/phoneNumber                                          controllers.CallForPaper.updatePhoneNumber()
 POST          /cfp/profile/phoneNumber/confirm                                   controllers.CallForPaper.confirmPhone()
-POST          /cfp/profile/notificationsPreferences                              controllers.CallForPaper.saveNotificationPreferences()
 
 # Email Digests
 GET           /cfp/backoffice/digests                                           controllers.Backoffice.showDigests()
@@ -384,3 +383,7 @@ GET           /cfpadmin/fixForInvalidSpeakers                                   
 
 # RSS Feed accepted talks
 GET           /rss/accepted                                                     controllers.RestAPI.RSSFeedAcceptedProposals()
+
+# Common preferences between GT & Comitee
+GET           /cfp/preferences                                                  controllers.CFPPreferencesController.index()
+POST          /cfp/preferences                                                  controllers.CFPPreferencesController.saveNotificationPreferences()
